fix: resolve circular import differently, so that module inspection can actually see the contents when called outside of a function
